RedstoneAI/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
amulet-core>=1.9.32
anthropic>=0.26.0
python-multipart>=0.0.9
jinja2>=3.1.0
